Customized notifications panel to Drawer from material ui, moved navbar and notifications to component folder, added normalizedString function to helpers
