+++ src/pages/Layout.jsx
@@ -1,5 +1,5 @@
 import { Footer } from "../components/Footer"
-import { CustomNavbar } from "../navbar/Navbar"
+import { CustomNavbar } from "../components/Navbar"
 import { Outlet } from "react-router-dom"
 
 export const Layout = () => {
